requirements.txt.txt
streamlit
gspread
oauth2client
psycopg2-binary
openpyxl
pandas
Pillow
requests
google-auth
pygsheets
matplotlib
seaborn
firebase-admin
streamlit-aggrid
plotly
SQLAlchemy
python-dotenv
redis
pyyaml
pymongo
requests-oauthlib
pycryptodome
